updated tests

+++ integration/correctness/src/test/resources/sql/test/cases-sinks-schema.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS events_wide_test
 (
     series_storage Int32,
     app Int32,
-    id Int32,
+    id UInt64,
     "timestamp" Float64,
     from Float64,
     to Float64,
@@ -12,7 +12,7 @@ CREATE TABLE IF NOT EXISTS events_narrow_test
 (
     series_storage Int32,
     app Int32,
-    id Int32,
+    id UInt64,
     "timestamp" Float64,
     from Float64,
     to Float64,
@@ -22,7 +22,7 @@ CREATE TABLE IF NOT EXISTS events_influx_test
 (
     series_storage Int32,
     app Int32,
-    id Int32,
+    id UInt64,
     "timestamp" Float64,
     from Float64,
     to Float64,
@@ -32,7 +32,7 @@ CREATE TABLE IF NOT EXISTS events_narrow_ivolga_test
 (
     series_storage Int32,
     app Int32,
-    id Int32,
+    id UInt64,
     "timestamp" Float64,
     from Float64,
     to Float64,
@@ -42,7 +42,7 @@ CREATE TABLE IF NOT EXISTS events_wide_ivolga_test
 (
     series_storage Int32,
     app Int32,
-    id Int32,
+    id UInt64,
     "timestamp" Float64,
     from Float64,
     to Float64,
